app/src/main/java/com/example/carolinereid/musicexchange/DoubleBass.java
package com.example.carolinereid.musicexchange;

import com.example.carolinereid.musicexchange.Behaviours.Playable;
import com.example.carolinereid.musicexchange.Behaviours.Sellable;

/**
 * Created by carolinereid on 27/10/2017.
 */

public class DoubleBass extends Instrument {
    private int numStrings;

    public DoubleBass(String material, String colour, String type, int costPrice, double retailPrice, int numStrings) {
        super(material, colour, type, costPrice, retailPrice);
        this.numStrings = numStrings;
    }

    public int getNumStrings() { return this.numStrings; }

    @Override
    public String makeSound() {
        return "mellow and deep";
    }
}
